draft2_nps.py
I identified a problem if we have several nps in one sentence, then there would be a mess, as they all will be appended together to one list. So, I suggested a way to have another layer of embedded lists. It is a bit messy too, as there are now three layers of embedded lists. Maybe we shall think of saving in texts instead of saving in lists at one point. Let's see what we could do today!