Update select_nps.py
Some info about update:
Changed dictionaries into lists, because based on the fact that we are going to deal with data, where people start learning the language, it could be that due to errors some token are duplicates. With dictionaries we would unintentionally overwrite those tokens. This could be avoided with token_IDs as keys, but I chose lists for now.